Rename `pkg/chain/ethereum` to `pkg/chain/ethereum_v1`
We need to make a transition towards the new
chain package structure required for V2. Here
is the first step.
