+++ cmd/start.go
@@ -3,6 +3,7 @@ package cmd
 import (
 	"context"
 	"fmt"
+	"github.com/keep-network/keep-core/pkg/chain/ethereum_v1"
 	"github.com/keep-network/keep-core/pkg/operator"
 	"time"
 
@@ -16,7 +17,6 @@ import (
 	"github.com/keep-network/keep-core/config"
 	"github.com/keep-network/keep-core/pkg/beacon"
 	"github.com/keep-network/keep-core/pkg/chain"
-	"github.com/keep-network/keep-core/pkg/chain/ethereum"
 	"github.com/keep-network/keep-core/pkg/firewall"
 	"github.com/keep-network/keep-core/pkg/net/libp2p"
 	"github.com/keep-network/keep-core/pkg/net/retransmission"
@@ -84,14 +84,14 @@ func Start(c *cli.Context) error {
 		)
 	}
 
-	operatorPrivateKey, operatorPublicKey, err := ethereum.ChainPrivateKeyToOperatorKeyPair(
+	operatorPrivateKey, operatorPublicKey, err := ethereum_v1.ChainPrivateKeyToOperatorKeyPair(
 		ethereumKey.PrivateKey,
 	)
 	if err != nil {
 		return err
 	}
 
-	chainProvider, err := ethereum.Connect(ctx, config.Ethereum)
+	chainProvider, err := ethereum_v1.Connect(ctx, config.Ethereum)
 	if err != nil {
 		return fmt.Errorf("error connecting to Ethereum node: [%v]", err)
 	}
