+++ pkg/chain/ethereum_v1/signer_test.go
@@ -1,4 +1,4 @@
-package ethereum
+package ethereum_v1
 
 import (
 	"crypto/rand"
